collecting points for creating rentings and losing when canceling, applying...

See merge request tim299642819/RentACarWebApp!28

+++ RentACarREST/src/dao/CustomerTypeDAO.java
@@ -57,7 +57,7 @@ public class CustomerTypeDAO {
 		}
 		return null;
 	}
-	public CustomerType getByPoints(Integer collectedPoints) {
+	public CustomerType getByPoints(Double collectedPoints) {
 		int maxPoints = 0;
 		CustomerType customerType = null;
 		for(CustomerType c : customerTypes) {
